requirements.txt
aliyun-python-sdk-core==2.3.5
aliyun-python-sdk-core-v3==2.3.7
aliyun-python-sdk-ecs==4.0.0
bcrypt==3.1.1
beautifulsoup4==4.6.0
blinker==1.4
cffi==1.9.1
crcmod==1.7
ecdsa==0.13
Flask==0.10.1
Flask-Cors==3.0.2
Flask-Login==0.3.0
Flask-Mail==0.9.1
flask-marshmallow==0.8.0
Flask-Principal==0.4.0
Flask-SocketIO==2.8.1
Flask-SQLAlchemy==2.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
marshmallow==2.13.5
marshmallow-sqlalchemy==0.13.1
oss2==2.3.3
paramiko==1.15.2
pep8==1.7.0
plumbum==1.5.0
pycparser==2.17
pycrypto==2.6.1
PyMySQL==0.7.4
python-engineio==1.1.0
python-gitlab==0.17
python-socketio==1.6.1
requests==2.12.4
six==1.10.0
SQLAlchemy==1.0.8
Werkzeug==0.11.11
